requirements.txt
fastapi
uvicorn
sentence-transformers
rank-bm25
pyahocorasick
torch
numpy
